code/python/packaging/requirements-packaging.txt
pyinstaller>=6.11,<7.0
pywebview>=5.3,<6.0
